requirements.txt
# Core ML
numpy
pandas
scikit-learn

# Vision
opencv-python
# albumentations (Desactive car non compatible Windows/Python 3.14 - Regle 4)

# Training
torch
torchvision
timm

# Export / Inference
onnx
onnxruntime

# Utilitaires et Visualisation
pyyaml
tqdm
matplotlib
seaborn
ipykernel
